+++ pom.xml
@@ -136,6 +136,12 @@ SOFTWARE.
       <version>0.2.1</version>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>wtf.g4s8</groupId>
+      <artifactId>matchers-json</artifactId>
+      <version>1.0.3</version>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
   <build>
     <testResources>
